Marking message serialization public.
We are utilizing other TcpStream implementations, which are not natively cloneable. However we can get a read-half and write-half which allows us to still construct a `Sender` and `Receiver` for the necessary components where the parts can be independently written to and read from concurrently.

In order to do this, we need to build a new `Sender` and `Receiver` which are based on a non-cloneable, but splittable stream (which we'll keep private in our usage). However we need to access the serialization calls from another crate, hence this change.